+++ custom_components/const.py
@@ -1,0 +1,11 @@
+API_ENDPOINT = 'https://api.weather.gov'
+USER_AGENT = 'Home Assistant'
+DEFAULT_ICON = 'mdi:alert'
+DEFAULT_NAME = 'NWS Alerts'
+CONF_ZONE_ID = 'zone_id'
+ZONE_ID = ''
+VERSION = '1.0'
+ISSUE_URL = 'https://github.com/finity69x2/nws_alert'
+DOMAIN = 'nws_alerts'
+PLATFORM = 'sensor'
+ATTRIBUTION = 'Data provided by Weather.gov'
