+++ custom_components/manifest.json
@@ -1,0 +1,9 @@
+{
+    "domain": "nws_alerts",
+    "name": "NWS Alerts",
+    "documentation": "https://github.com/moralmunky/Home-Assistant-Mail-And-Packages",
+    "dependencies": [],
+    "codeowners": ["@finity69x2"],
+    "config_flow": true,
+    "requirements": []
+  }
