+++ custom_components/translations/en.json
@@ -1,0 +1,31 @@
+{
+    "config": {
+        "abort": {
+            "single_instance_allowed": "Only a single configuration of Mail and Packages is allowed."
+        },        
+        "step": {
+            "user": {
+                "data": {
+                    "name": "Friendly Name",
+                    "zone_id": "Zone ID(s)"
+                 },
+                "description": "You can find your Zone or County ID by going to https://alerts.weather.gov/, scroll down to your state and click on the “zone list” and/or “county list” then look for the entry for your county.\nSeperate multiple zones with commas ie: PAC049,WVC031",
+                "title": "NWS Alerts"
+            }
+        },
+        "title": "NWS Alerts"
+    },
+    "options": {
+        "step": {
+            "init": {
+                "data": {
+                    "name": "Friendly Name",
+                    "zone_id": "Zone ID(s)"
+                },
+                "description": "You can find your Zone or County ID by going to https://alerts.weather.gov/, scroll down to your state and click on the “zone list” and/or “county list” then look for the entry for your county.\nSeperate multiple zones with commas ie: PAC049,WVC031",
+                "title": "NWS Alerts"
+            }
+        },
+        "title": "NWS Alerts"
+    }
+}
