+++ hacs.json
@@ -1,0 +1,6 @@
+{
+    "name": "NWS Alerts",
+    "domains": [ "sensor" ],
+    "homeassistant": "0.95.4",
+    "iot_class": "Cloud Polling"
+}
